Change `analytics` to `statistics`
We don't have an `analytics` permission type (this is `statistics`). Changing this
caused an issue with the tests, so for some reason we have to make the browser
wait before clicking our settings. This won't be an issue in real life, and I've already
spent ages on this, so getting it good enough for now. I've also taken the
opportunity to refactor how we set the cookies, so it's all done in one go, which
came out during a debug.